+++ frontend/src/modules/index.ts
@@ -2,7 +2,7 @@
  * @summary All UI components & routes of the app resides here
  *
  */
-export * from './alarms/AlarmsPage';
+export * from './alarms/limits/AlarmLimitsPage';
 export * from './app';
 export * from './controllers';
 export * from './displays';
